+++ changelog.md
@@ -1,3 +1,9 @@
+# v0.0.1-alpha.2
+
+- Macro Popup (Including button to open popup in PauseLayer)
+- Repository Link
+- No more clicking on the 90th frame :3
+
 # v0.0.1-alpha.1
 
 - Frame Label